LogWatch/requirements.txt
confluent_kafka==2.10.0
fastapi==0.115.12
mysql-connector-python
pandas==2.3.0
pydantic==2.11.5
python-dotenv==1.1.0
bcrypt
redis==6.2.0
Requests==2.32.3
streamlit==1.45.1
uvicorn
